feat(list): support sorting and reordering user lists
