+++ projects/client/src/lib/requests/models/UserListsSortBy.ts
@@ -1,0 +1,1 @@
+export type UserListsSortBy = 'rank' | 'name' | 'updated_at' | 'created_at';
