+++ projects/client/src/lib/sections/lists/user/PersonalListsPaginated.svelte
@@ -1,12 +1,24 @@
 <script lang="ts">
   import PaginatedList from "$lib/components/lists/PaginatedList.svelte";
   import { useDiscover } from "$lib/features/filters/useDiscover";
+  import type { UserListsSortBy } from "$lib/requests/models/UserListsSortBy.ts";
   import { DEFAULT_LISTS_DRILL_SIZE } from "$lib/utils/constants";
   import type { PersonalListType } from "./models/PersonalListType";
+  import type { SortDirection } from "./models/SortDirection";
   import { usePersonalListsSummary } from "./usePersonalListsSummary";
   import UserList from "./UserList.svelte";
 
-  const { slug, type }: { slug: string; type: PersonalListType } = $props();
+  const {
+    slug,
+    type,
+    sortBy,
+    sortHow,
+  }: {
+    slug: string;
+    type: PersonalListType;
+    sortBy?: UserListsSortBy | Nil;
+    sortHow?: SortDirection | Nil;
+  } = $props();
 
   const { mode } = useDiscover();
 </script>
@@ -19,6 +31,8 @@
         ...params,
         slug,
         limit: DEFAULT_LISTS_DRILL_SIZE,
+        sortBy,
+        sortHow,
       })}
   >
     {#snippet items(items)}
